% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.2DII-package.R
\docType{package}
\name{Data.2DII-package}
\alias{Data.2DII}
\alias{Data.2DII-package}
\title{Data.2DII: Data package captured by 2DII}
\description{
The package contains the dataset used in 2DII reports.
}
\author{
\strong{Maintainer}: Ana Rivera \email{ana@2degrees-investing.org} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
