% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESG_france_2021.R
\docType{data}
\name{ESG_france_2021}
\alias{ESG_france_2021}
\title{ESG preferences survey France 2021}
\format{
A tibble with 1534 rows and 342 variables:
\describe{
\item{record}{ Record number}
\item{uuid}{ Participant identifier}
\item{date}{ Completion time and date}
\item{status}{ Participant status}
\item{HIDCOUNTRY}{ Hidden for country.}
\item{Hid_IMAGE_FRAr1}{https://surveys.globaltestmarket.com/survey/selfserve/1ab6/112180770/Q1x8.png}
\item{Hid_IMAGE_FRAr2}{https://surveys.globaltestmarket.com/survey/selfserve/1ab6/112180770/FR_Section2.png}
\item{Hid_IMAGE_FRAr3}{https://surveys.globaltestmarket.com/survey/selfserve/1ab6/112180770/FR_Q2Bx9.png}
\item{Hid_IMAGE_FRAr4}{https://surveys.globaltestmarket.com/survey/selfserve/1ab6/112180770/FR_Q2Bx10_1.PNG}
\item{Hid_IMAGE_FRAr5}{https://surveys.globaltestmarket.com/survey/selfserve/1ab6/112180770/FR_Q2Bx10_2.PNG}
\item{Hid_IMAGE_FRAr6}{https://surveys.globaltestmarket.com/survey/selfserve/1ab6/112180770/FR_Q2Bx10_3.PNG}
\item{Hid_IMAGE_FRAr7}{https://surveys.globaltestmarket.com/survey/selfserve/1ab6/112180770/FR_Q2Bx10_4.PNG}
\item{Hid_IMAGE_FRAr8}{https://surveys.globaltestmarket.com/survey/selfserve/1ab6/112180770/FR_Investment1.PNG}
\item{Hid_IMAGE_FRAr9}{https://surveys.globaltestmarket.com/survey/selfserve/1ab6/112180770/FR_Investment2.PNG}
\item{Hid_IMAGE_FRAr10}{https://surveys.globaltestmarket.com/survey/selfserve/1ab6/112180770/FR_Investment3.PNG}
\item{Hid_IMAGE_FRAr11}{https://surveys.globaltestmarket.com/survey/selfserve/1ab6/112180770/FR_Investment4.PNG}
\item{Hid_IMAGE_FRAr12}{https://surveys.globaltestmarket.com/survey/selfserve/1ab6/112180770/FR_Investment5.PNG}
\item{Hid_IMAGE_IREr1}{https://surveys.globaltestmarket.com/survey/selfserve/1ab6/112180770/Q1x8.png}
\item{Hid_IMAGE_IREr2}{https://surveys.globaltestmarket.com/survey/selfserve/1ab6/112180770/Section2.png}
\item{Hid_IMAGE_IREr3}{https://surveys.globaltestmarket.com/survey/selfserve/1ab6/112180770/Q2Bx9.png}
\item{Hid_IMAGE_IREr4}{https://surveys.globaltestmarket.com/survey/selfserve/1ab6/112180770/Q2Bx10_1.PNG}
\item{Hid_IMAGE_IREr5}{https://surveys.globaltestmarket.com/survey/selfserve/1ab6/112180770/Q2Bx10_2.PNG}
\item{Hid_IMAGE_IREr6}{https://surveys.globaltestmarket.com/survey/selfserve/1ab6/112180770/Q2Bx10_3.PNG}
\item{Hid_IMAGE_IREr7}{https://surveys.globaltestmarket.com/survey/selfserve/1ab6/112180770/Q2Bx10_4.PNG}
\item{Hid_IMAGE_IREr8}{https://surveys.globaltestmarket.com/survey/selfserve/1ab6/112180770/Investment1.PNG}
\item{Hid_IMAGE_IREr9}{https://surveys.globaltestmarket.com/survey/selfserve/1ab6/112180770/Investment2.PNG}
\item{Hid_IMAGE_IREr10}{https://surveys.globaltestmarket.com/survey/selfserve/1ab6/112180770/Investment3.PNG}
\item{Hid_IMAGE_IREr11}{https://surveys.globaltestmarket.com/survey/selfserve/1ab6/112180770/Investment4.PNG}
\item{Hid_IMAGE_IREr12}{https://surveys.globaltestmarket.com/survey/selfserve/1ab6/112180770/Investment5.PNG}
\item{Final_IMAGEr1}{Final_IMAGEr1}
\item{Final_IMAGEr2}{Final_IMAGEr2}
\item{Final_IMAGEr3}{Final_IMAGEr3}
\item{Final_IMAGEr4}{Final_IMAGEr4}
\item{Final_IMAGEr5}{Final_IMAGEr5}
\item{Final_IMAGEr6}{Final_IMAGEr6}
\item{Final_IMAGEr7}{Final_IMAGEr7}
\item{Final_IMAGEr8}{Final_IMAGEr8}
\item{Final_IMAGEr9}{Final_IMAGEr9}
\item{Final_IMAGEr10}{Final_IMAGEr10}
\item{Final_IMAGEr11}{Final_IMAGEr11}
\item{Final_IMAGEr12}{Final_IMAGEr12}
\item{Q1x1}{ What is your gender?}
\item{Q1x2r1}{ How old are you?}
\item{HIDAGE}{ Hidden recode for Q1x2}
\item{Q1x3}{ What is your education level?}
\item{Q1x3r6oe}{ What is your education level? - Other professional qualification}
\item{Q1x4_FR_IT_ES_BE_NL_DE_GR_EE_AT_IE}{ What is your average monthly household net income?}
\item{Q1x4_ES}{ What is your average monthly household net income?}
\item{Q1x4_BE}{ What is your average monthly household net income?}
\item{Q1x4_CZ}{ What is your average monthly household net income?}
\item{Q1x4_SE}{ What is your average monthly household net income?}
\item{Q1x4_PL}{ What is your average monthly household net income?}
\item{Q1x4_RO}{ What is your average monthly household net income?}
\item{Q1x4_DK}{ What is your average monthly household net income?}
\item{Q1x5_FR_IT_ES_BE_NL_DE_GR_EE_AT_IE}{ How much do you save monthly?}
\item{Q1x5_CZ}{ How much do you save monthly?}
\item{Q1x5_SE}{ How much do you save monthly?}
\item{Q1x5_PL}{ How much do you save monthly?}
\item{Q1x5_RO}{ How much do you save monthly?}
\item{Q1x5_DK}{ How much do you save monthly?}
\item{Q1x6}{ Do you currently have any financial investments such as investment funds, shares, life insurance etc.?}
\item{Q1x7}{ Are any of your financial investments in sustainable products?}
\item{Q1x8r1}{A - the substance of the invested money is preserved with very low risk}
\item{Q1x8r2}{B - gains above the interest rate are achieved with higher risk}
\item{Q1x8r3}{C - return oriented portfolio with balanced risk}
\item{Q1x8r4}{D - dynamic value growth from high gains but high risks}
\item{Q1x9r1}{Choosing how I invest is an important decision for me}
\item{Q1x9r2}{I would choose how I invest carefully}
\item{Q1x9r3}{What financial product I choose matters a lot to me}
\item{Q1x10r1}{Saving for retirement}
\item{Q1x10r2}{Long-term increase in assets}
\item{Q1x10r3}{Generate additional income}
\item{Q1x10r4}{I want to save money for children or relatives}
\item{Q1x10r5}{Something different}
\item{Q1x10x2}{ You have mentioned for your financial goals that you want to achieve something different.Could you please specify?}
\item{Q1x11r1}{My knowledge of finance and investment is good.}
\item{Q1x11r2}{My knowledge of sustainable finance is good.}
\item{CONSENT}{ The next question of this survey is about your political views, which is considered as sensitive data. It will be used by our client for data classification purpose only. It will remain confidential in line with our privacy policy. If answering the next question makes you uncomfortable, please feel free to choose the answer "I would prefer not to respond". Do you agree to answer this question on this basis?}
\item{Q1x14xFR}{ For which party would you vote if there would be national elections next week?}
\item{Q1x14xIT}{ For which party would you vote if there would be national elections next week?}
\item{Q1x14xES}{ For which party would you vote if there would be national elections next week?}
\item{Q1x14xBE}{ For which party would you vote if there would be national elections next week?}
\item{Q1x14xNL}{ For which party would you vote if there would be national elections next week?}
\item{Q1x14xIE}{ For which party would you vote if there would be national elections next week?}
\item{Q1x14xSE}{ How much do you save monthly?}
\item{Q1x14xPO}{ How much do you save monthly?}
\item{h_ctype}{ Hidden for defining respodent type}
\item{h_SC}{ Hidden for SC terminate position for STARTER.}
\item{HIDCELL}{ Hidden variable to ALLOCATE RANDOMLY WITH A SAM SPLIT ON GENDER AND HIDAGE}
\item{HIDVideoQ2Ax0}{ Hidden variable to ALLOCATE the video ID.}
\item{Q2Ax0play_count}{Play Count}
\item{Q2Ax0time_elapsed}{Elapsed Time}
\item{TimexQ2Ax0r1}{ VAR_TIME}
\item{Q2Ax1x1}{ You may want to invest in a way that changes company behaviour in the real economy, such as reducing CO2 emissions or improving working conditions in the supply chain. However, this impact is hard to prove. To do so, impact products might have an active investment strategy which leads to higher product fees or they are invested in young fast-growing companies which leads to higher potential risks as well as returns. As this type of investor, your objective is to have an impact in the real economy regardless the industry of the companies or of impact on financial performance. "I want my money to be invested in a way that impacts company behaviour in line with my impact objectives."}
\item{Q2Ax1x2}{ You may want to invest in a way that changes company behaviour in the real economy, such as reducing CO2 emissions or improving working conditions in the supply chain. Without private investments it won´t be possible to solve these sustainability issues otherwise. "I want my money to be invested in a way that impacts company behaviour in line with my impact objectives."}
\item{Q2Ax2}{ You may want to invest in a way that avoids investing in certain economic activities, such as weapons, coal, or alcohol because of moral, political, or religious reasons. The objective is to disassociate from these activities regardless of any impact on company behaviour or on financial performance. "I want my money to be invested in a way that excludes activities which do not correspond to my values.”}
\item{Q2Ax3}{ You may want to invest in a way that takes account of social, environmental and governance factors with the objective to improve financial performance. Including these factors can reduce the risk of investing in companies which might face reputational damages or miss market opportunities due to non-sustainable behaviour. The objective is to take account of social, environmental and governance factors regardless of any impact on company behaviour or certain industries/activities to avoid within your investing. "I want my money to be invested in such a way that social, environmental and governance risk and opportunity factors are given special consideration, because I believe you can earn more money that way."}
\item{Q2Ax4}{ You may want to invest in a way that takes no special considerations of social, environmental and governance factors. The objective is to maximizes your returns without the consideration of sustainability regardless of any impact on company behaviour or certain industries/activities to avoid within your investing. “I want my money invested in way that maximizes my returns without the consideration of sustainability.”}
\item{hidxQ2Ax5r1}{"I want my money to be invested in a way that impacts company behaviour in line with my impact objectives."}
\item{hidxQ2Ax5r2}{"I want my money to be invested in a way that excludes activities which do not correspond to my values."}
\item{hidxQ2Ax5r3}{"I want my money to be invested in such a way that social, environmental and governance risk and opportunity factors are given special consideration, because I believe you can earn more money that way."}
\item{hidxQ2Ax5r4}{"I want my money invested in way that maximizes my returns."}
\item{hidxQ2AxSelectionr1}{Saving for retirement}
\item{hidxQ2AxSelectionr2}{Long-term increase in assets}
\item{hidxQ2AxSelectionr3}{Generate additional income}
\item{hidxQ2AxSelectionr4}{I want to save money for children or relatives}
\item{hidxQ2AxSelectionr5}{$ {Q1x10x2.val}}
\item{hidxQ2Ax5pipe}{ Hidden for defining respodent type}
\item{Q2Ax5r1}{"I want my money to be invested in a way that impacts company behaviour in line with my impact objectives."}
\item{Q2Ax5r2}{"I want my money to be invested in a way that excludes activities which do not correspond to my values."}
\item{Q2Ax5r3}{"I want my money to be invested in such a way that social, environmental and governance risk and opportunity factors are given special consideration, because I believe you can earn more money that way."}
\item{Q2Ax5r4}{"I want my money invested in way that maximizes my returns."}
\item{HIDVideoQ2Bx0}{ Hidden variable to ALLOCATE the video ID.}
\item{Q2Bx0play_count}{Play Count}
\item{Q2Bx0time_elapsed}{Elapsed Time}
\item{TimexQ2Bx0r1}{ VAR_TIME}
\item{Q2Bx1r1}{Reduce poverty}
\item{Q2Bx1r2}{Reduce hunger}
\item{Q2Bx1r3}{Promote health}
\item{Q2Bx1r4}{Promote education}
\item{Q2Bx1r5}{Promote gender equality}
\item{Q2Bx1r6}{Promote clean drinking water and sanitation worldwide}
\item{Q2Bx1r7}{Promote renewable energy}
\item{Q2Bx1r8}{Promoting sustainable economic growth and decent work}
\item{Q2Bx1r9}{Promote sustainable infrastructure and innovation}
\item{Q2Bx1r10}{Reduce inequality}
\item{Q2Bx1r11}{Promoting sustainable construction and refurbishment}
\item{Q2Bx1r12}{Promoting sustainable products}
\item{Q2Bx1r13}{Promoting climate protection}
\item{Q2Bx1r14}{Improving water quality and fish stocks}
\item{Q2Bx1r15}{Improving nature conservation and biodiversity}
\item{Q2Bx1r16}{Promote peace, justice, and the strong institutions}
\item{Q2Bx1r17}{Promoting partnerships for sustainable development}
\item{Q2Bx2}{ Additional to the issues mentioned in the last question, are there additional social, environmental or governance issues which are important to you?}
\item{Q2Bx3}{ Executing voting rights at shareholder meetings is one strategy that gives a good probability to have impact on sustainability issues. As an investor, how often would you like to be asked about your opinion how the investment firm should vote in those meetings?}
\item{Q2Bx4}{ A financial product may have a certain likelihood to achieve real-world impact on the sustainability issues you have identified, depending on its investment strategy and how it is implemented. How much evidence do you want to see on the real world impact of your investments?}
\item{Q2Bx4r6oe}{ A financial product may have a certain likelihood to achieve real-world impact on the sustainability issues you have identified, depending on its investment strategy and how it is implemented. How much evidence do you want to see on the real world impact of your investments? - I want something else}
\item{hidxPipeQ2Bx5}{ Hidden for 2B.5-2B.6-2B.7}
\item{hidxPipeQ2Bx8}{ Hidden for piping at 2B.8.}
\item{Q2Bx5r1}{Oil}
\item{Q2Bx5r2}{Gas}
\item{Q2Bx5r3}{Coal}
\item{Q2Bx5r4}{Nuclear energy}
\item{Q2Bx5r5}{Palm oil}
\item{Q2Bx5r6}{Genetically modified organisms}
\item{Q2Bx5r7}{Sale of pesticides and biocides}
\item{Q2Bx5r8}{Animal testing}
\item{Q2Bx5r9}{Environmental standards violations in the supply chain}
\item{Q2Bx6r1}{Human rights violations in the supply chain}
\item{Q2Bx6r2}{Labour rights violations in the supply chain}
\item{Q2Bx6r3}{Weapons and/or armaments}
\item{Q2Bx7r1}{Tobacco}
\item{Q2Bx7r2}{Alcohol}
\item{Q2Bx7r3}{Pork-based products}
\item{Q2Bx7r4}{Research involving human embryos}
\item{Q2Bx7r5}{Gambling}
\item{Q2Bx7r6}{Pornography}
\item{Q2Bx7r7}{Abortion}
\item{Q2Bx7r8}{Contraceptives}
\item{Q2Bx7r9}{Violent video games}
\item{Q2Bx7r10}{Cannabis}
\item{Q2Bx8r1}{Violation of the United Nations Global Compact principles (UN principles on human rights, labour, environment and ant-corruption)}
\item{Q2Bx8r2}{Companies without women in the management}
\item{Q2Bx8r3}{Non-disclosure of directors' salaries}
\item{Q2Bx8r4}{Controversies in the field of corruption}
\item{Q2Bx8r5}{Tax avoidance strategies and identified infringements}
\item{Q2Bx8r6}{Anti-Competitive Behaviour}
\item{Q2Bx8r7}{Accounting fraud}
\item{Q2Bx9}{ If you want to exclude/avoid companies which generate revenues with the activities which are important to you, how strict do you want the exclusion to be?}
\item{Q2Bx10}{ You want to invest in a way that takes account of social, environmental and governance factors with the objective to improve financial performance. There are different strategies to reduce risks or increase opportunities related to social, environmental or governance factors. Please select the options which you find most promising to achieve this.}
\item{Q2Cx1}{ You have decided to invest sustainably. If you cannot find a financial product that matches the exact sustainability objectives you have identified, what would best describe your reaction?}
\item{Q2Cx2}{ Please rate the following statement}
\item{HID_VALUE_X_FR_IT_ES_BE_NL_DE_GR_EE_AT_IE}{ Hidden for HID_VALUE_X}
\item{HID_VALUE_Y_FR_IT_ES_BE_NL_DE_GR_EE_AT_IE}{ Hidden for HID_VALUE_Y}
\item{HID_VALUE_Z_FR_IT_ES_BE_NL_DE_GR_EE_AT_IE}{ Hidden for HID_VALUE_Z}
\item{HID_VALUE_X_CZ}{ Hidden for HID_VALUE_X}
\item{HID_VALUE_Y_CZ}{ Hidden for HID_VALUE_Y}
\item{HID_VALUE_Z_CZ}{ Hidden for HID_VALUE_Z}
\item{HID_VALUE_X_SE}{ Hidden for HID_VALUE_X}
\item{HID_VALUE_Y_SE}{ Hidden for HID_VALUE_Y}
\item{HID_VALUE_Z_SE}{ Hidden for HID_VALUE_Z}
\item{HID_VALUE_X_PL}{ Hidden for HID_VALUE_X}
\item{HID_VALUE_Y_PL}{ Hidden for HID_VALUE_Y}
\item{HID_VALUE_Z_PL}{ Hidden for HID_VALUE_Z}
\item{HID_VALUE_X_RO}{ Hidden for HID_VALUE_X}
\item{HID_VALUE_Y_RO}{ Hidden for HID_VALUE_Y}
\item{HID_VALUE_Z_RO}{ Hidden for HID_VALUE_Z}
\item{HID_VALUE_X_DK}{ Hidden for HID_VALUE_X}
\item{HID_VALUE_Y_DK}{ Hidden for HID_VALUE_Y}
\item{HID_VALUE_Z_DK}{ Hidden for HID_VALUE_Z}
\item{Pipe_VALUE_Xr1}{ Hidden for Q2Cx3xA and Q2Cx3xB}
\item{Pipe_VALUE_Yr1}{ Hidden for Q2Cx3xA and Q2Cx3xB}
\item{Pipe_VALUE_Zr1}{ Hidden for Q2Cx3xA and Q2Cx3xB}
\item{Q2Cx3xB}{ By accepting a financial sacrifice up to a certain amount of your savings, you earn not only a financial return, but also deliver true impact and get evidence that your objectives are achieved in the real economy. There are estimates that this evidence is stronger, the higher the financial sacrifice. In the slider below, you can choose between 0 and 20\% of your final savings. Given your saving behaviour, you would save on average after 10 years with 4\% interest. In this case, 1\% sacrifice would mean giving up on and 20\% would mean giving up for achieving impact in the real economy. Please select how much you would be willing to sacrifice at most}
\item{Q2Cx3xA}{ Generating impact on your sustainability objectives in the real economy is hard to prove. The most evidence for impact might be expected from active management strategies. In the slider below, you can choose between 0 and 20\% of your final savings. Given your saving behaviour, you would save on average after 10 years with 4\% interest. In this case, 1\% sacrifice would mean giving up on and 20\% would mean giving up for an active management strategy. Please select how much you would be willing to sacrifice at most}
\item{Q3x1r1}{Boring}
\item{Q3x1r2}{Interesting}
\item{Q3x1r3}{Complex}
\item{Q3x1r4}{Easy}
\item{Q3x1r5}{Misleading}
\item{Q3x1r6}{Informative}
\item{Q3x1r7}{Unreliable}
\item{Q3x1r8}{Trustworthy}
\item{Q3x1r9}{Convincing}
\item{Q3x1r10}{Difficult}
\item{Q3x1r11}{Bad}
\item{Q3x1r12}{Good}
\item{Q3x2}{ Which of the following sentences best describes your understanding of the environmental characteristics associated with this financial product?}
\item{Q3x3}{ Actually, the Equity Fund is invested in companies that pollute less than their competitors. The Equity Fund calculates the difference between the CO2 emissions of these companies and the average of the market every year. There is no evidence that investing more in the Equity Fund will change the operations of these companies though. So it cannot be said that “the more money invested in the fund, the more CO2 reduction activities are developed, the greater your environmental impact.” In the light of this explanation, how would you characterize the marketing claim for this financial product?}
\item{Q3x7r1}{Boring}
\item{Q3x7r2}{Interesting}
\item{Q3x7r3}{Complex}
\item{Q3x7r4}{Easy}
\item{Q3x7r5}{Misleading}
\item{Q3x7r6}{Informative}
\item{Q3x7r7}{Unreliable}
\item{Q3x7r8}{Trustworthy}
\item{Q3x7r9}{Convincing}
\item{Q3x7r10}{Difficult}
\item{Q3x7r11}{Bad}
\item{Q3x7r12}{Good}
\item{Q3x8}{ Which of the following sentences best describes your understanding of the environmental characteristics associated with this financial product?}
\item{Q3x9}{ Actually, the Green Equity Fund is invested in the stocks of companies whose production plans are aligned with a 2 °C climate transition pathway. There is no evidence that investing in this fund will contribute to lowering companies’ emissions though. So, it cannot be said that “the more money invested in the fund, the more CO2 emissions are reduced in the world, the lower the increase in global temperatures.” In the light of this explanation, how would you characterize the marketing claim for this financial product?}
\item{Q3x10r1}{Boring}
\item{Q3x10r2}{Interesting}
\item{Q3x10r3}{Complex}
\item{Q3x10r4}{Easy}
\item{Q3x10r5}{Misleading}
\item{Q3x10r6}{Informative}
\item{Q3x10r7}{Unreliable}
\item{Q3x10r8}{Trustworthy}
\item{Q3x10r9}{Convincing}
\item{Q3x10r10}{Difficult}
\item{Q3x10r11}{Bad}
\item{Q3x10r12}{Good}
\item{Q3x11}{ Which of the following sentences best describes your understanding of the environmental characteristics associated with this financial product?}
\item{Q3x12}{ Actually, the Green Bond Fund is invested is in the bonds of companies regardless of the nature of their activities (e.g. active in coal-power plants, gas-guzzling cars as well as renewable power, electric car). The Green Bond Fund only invests in the “green” element of these bonds, and the amount available in this green element matches the total amount of money that the companies have invested in environmentally-friendly activities. There is no evidence that your money will increase the number of environmentally-friendly activities though. So it cannot be said that “the more money invested in the fund, the more CO2 reduction activities are developed, the greater your environmental impact.” In the light of this explanation, how would you characterize the marketing claim for this financial product?}
\item{Q3x13r1}{Boring}
\item{Q3x13r2}{Interesting}
\item{Q3x13r3}{Complex}
\item{Q3x13r4}{Easy}
\item{Q3x13r5}{Misleading}
\item{Q3x13r6}{Informative}
\item{Q3x13r7}{Unreliable}
\item{Q3x13r8}{Trustworthy}
\item{Q3x13r9}{Convincing}
\item{Q3x13r10}{Difficult}
\item{Q3x13r11}{Bad}
\item{Q3x13r12}{Good}
\item{Q3x14}{ It is not, to date, possible to measure the impact of a financial product on CO2 emissions. All claims containing emission figures are thus misleading. The above claim is an example of a non-misleading claim}
\item{Q4x1r1}{Investing in other activities beside EEI/SEI for risk diversification is ok for me.}
\item{Q4x1r2}{EEI/SEI products should not invest in activities that are harmful to the environment.}
\item{Q4x1r3}{Products labelled EEI/SEI should not invest in any other activities.}
\item{Q4x1r4}{I would accept higher risks/lower returns if I could invest in products that facilitate only EE/SE.}
\item{Q4x3r1}{Investment 1}
\item{Q4x3r2}{Investment 2}
\item{Q4x3r3}{Investment 3}
\item{Q4x3r4}{Investment 4}
\item{Q4x3r5}{Investment 5}
\item{pagetimeInfo}{(Info) Comment element}
\item{pagetimeQ1x1}{(Q1x1) What is your gender?}
\item{pagetimeQ1x2}{(Q1x2) How old are you?}
\item{pagetimeQ1x3}{(Q1x3) What is your education level?}
\item{pagetimeQ1x4_FR_IT_ES_BE_NL_DE_GR_EE_AT_IE}{(Q1x4_FR_IT_ES_BE_NL_DE_GR_EE_AT_IE) What is your average monthly household net income?}
\item{pagetimeQ1x4_ES}{(Q1x4_ES) What is your average monthly household net income?}
\item{pagetimeQ1x4_BE}{(Q1x4_BE) What is your average monthly household net income?}
\item{pagetimeQ1x4_CZ}{(Q1x4_CZ) What is your average monthly household net income?}
\item{pagetimeQ1x4_SE}{(Q1x4_SE) What is your average monthly household net income?}
\item{pagetimeQ1x4_PL}{(Q1x4_PL) What is your average monthly household net income?}
\item{pagetimeQ1x4_RO}{(Q1x4_RO) What is your average monthly household net income?}
\item{pagetimeQ1x4_DK}{(Q1x4_DK) What is your average monthly household net income?}
\item{pagetimeQ1x5_FR_IT_ES_BE_NL_DE_GR_EE_AT_IE}{(Q1x5_FR_IT_ES_BE_NL_DE_GR_EE_AT_IE) How much do you save monthly?}
\item{pagetimeQ1x5_CZ}{(Q1x5_CZ) How much do you save monthly?}
\item{pagetimeQ1x5_SE}{(Q1x5_SE) How much do you save monthly?}
\item{pagetimeQ1x5_PL}{(Q1x5_PL) How much do you save monthly?}
\item{pagetimeQ1x5_RO}{(Q1x5_RO) How much do you save monthly?}
\item{pagetimeQ1x5_DK}{(Q1x5_DK) How much do you save monthly?}
\item{pagetimeQ1x6}{(Q1x6) Do you currently have any financial investments such as investment funds, shares, life insurance etc.?}
\item{pagetimeQ1x7}{(Q1x7) Are any of your financial investments in sustainable products?}
\item{pagetimeQ1x8}{(Q1x8) In the following graph you can see the potential gains (in green) and the potential losses (in red) compared to your input out of an investment strategy. Losses and gains occur equally often. Please click on image to enlarge. Which of the presented risk categories corresponds most closely to your preferred allocation for your money?}
\item{pagetimeQ1x9}{(Q1x9) Please rate the following statements:}
\item{pagetimeQ1x10}{(Q1x10) What are your financial goals, thus what do you want to achieve financially with your money?}
\item{pagetimeQ1x10x2}{(Q1x10x2) You have mentioned for your financial goals that you want to achieve something different.Could you please specify?}
\item{pagetimeQ1x11}{(Q1x11) Please rate these statements:}
\item{pagetimeCONSENT}{(CONSENT) The next question of this survey is about your political views, which is considered as sensitive data. It will be used by our client for data classification purpose only. It will remain confidential in line with our privacy policy. If answering the next question makes you uncomfortable, please feel free to choose the answer "I would prefer not to respond". Do you agree to answer this question on this basis?}
\item{pagetimeQ1x14xFR}{(Q1x14xFR) For which party would you vote if there would be national elections next week?}
\item{pagetimeQ1x14xIT}{(Q1x14xIT) For which party would you vote if there would be national elections next week?}
\item{pagetimeQ1x14xES}{(Q1x14xES) For which party would you vote if there would be national elections next week?}
\item{pagetimeQ1x14xBE}{(Q1x14xBE) For which party would you vote if there would be national elections next week?}
\item{pagetimeQ1x14xNL}{(Q1x14xNL) For which party would you vote if there would be national elections next week?}
\item{pagetimeQ1x14xIE}{(Q1x14xIE) For which party would you vote if there would be national elections next week?}
\item{pagetimeQ1x14xSE}{(Q1x14xSE) How much do you save monthly?}
\item{pagetimeQ1x14xPO}{(Q1x14xPO) How much do you save monthly?}
\item{pagetimeSection2}{(Section2) Comment element}
\item{pagetimeQ2Ax0}{(Q2Ax0) We will now show you a video about your financial and non-financial goals.}
\item{pagetimeQ2Ax1x1}{(Q2Ax1x1) You may want to invest in a way that changes company behaviour in the real economy, such as reducing CO2 emissions or improving working conditions in the supply chain. However, this impact is hard to prove. To do so, impact products might have an active investment strategy which leads to higher product fees or they are invested in young fast-growing companies which leads to higher potential risks as well as returns. As this type of investor, your objective is to have an impact in the real economy regardless the industry of the companies or of impact on financial performance. "I want my money to be invested in a way that impacts company behaviour in line with my impact objectives."}
\item{pagetimeQ2Ax1x2}{(Q2Ax1x2) You may want to invest in a way that changes company behaviour in the real economy, such as reducing CO2 emissions or improving working conditions in the supply chain. Without private investments it won´t be possible to solve these sustainability issues otherwise. "I want my money to be invested in a way that impacts company behaviour in line with my impact objectives."}
\item{pagetimeQ2Ax2}{(Q2Ax2) You may want to invest in a way that avoids investing in certain economic activities, such as weapons, coal, or alcohol because of moral, political, or religious reasons. The objective is to disassociate from these activities regardless of any impact on company behaviour or on financial performance. "I want my money to be invested in a way that excludes activities which do not correspond to my values.”}
\item{pagetimeQ2Ax3}{(Q2Ax3) You may want to invest in a way that takes account of social, environmental and governance factors with the objective to improve financial performance. Including these factors can reduce the risk of investing in companies which might face reputational damages or miss market opportunities due to non-sustainable behaviour. The objective is to take account of social, environmental and governance factors regardless of any impact on company behaviour or certain industries/activities to avoid within your investing. "I want my money to be invested in such a way that social, environmental and governance risk and opportunity factors are given special consideration, because I believe you can earn more money that way."}
\item{pagetimeQ2Ax4}{(Q2Ax4) You may want to invest in a way that takes no special considerations of social, environmental and governance factors. The objective is to maximizes your returns without the consideration of sustainability regardless of any impact on company behaviour or certain industries/activities to avoid within your investing. “I want my money invested in way that maximizes my returns without the consideration of sustainability.”}
\item{pagetimeQ2Ax5}{(Q2Ax5) You have chosen more than one sustainability objective. Please rank their priority in the order you would like to have them implemented for your primary financial goal: \link{pipe: hidxQ2Ax5pipe}}
\item{pagetimeQ2Bx0}{(Q2Bx0) We will now show you a second video about your sustainable development goals.}
\item{pagetimeQ2Bx1}{(Q2Bx1) Investors often own shares in companies and can influence how these companies act. Imagine that you could decide alone about how a company should act. Which of these sustainability issues would be most important to you in this position?}
\item{pagetimeQ2Bx2}{(Q2Bx2) Additional to the issues mentioned in the last question, are there additional social, environmental or governance issues which are important to you?}
\item{pagetimeQ2Bx3}{(Q2Bx3) Executing voting rights at shareholder meetings is one strategy that gives a good probability to have impact on sustainability issues. As an investor, how often would you like to be asked about your opinion how the investment firm should vote in those meetings?}
\item{pagetimeQ2Bx4}{(Q2Bx4) A financial product may have a certain likelihood to achieve real-world impact on the sustainability issues you have identified, depending on its investment strategy and how it is implemented. How much evidence do you want to see on the real world impact of your investments?}
\item{pagetimeQ2Bx5}{(Q2Bx5) If you could exclude companies who are involved in certain environmental activities from your investments, how important would it be to you to exclude the following activities\link{pipe: hidxPipeQ2Bx5}?}
\item{pagetimeQ2Bx6}{(Q2Bx6) If you could exclude companies who are involved in certain social activities from your investments, how important would it be to you to exclude the following activities\link{pipe: hidxPipeQ2Bx5}?}
\item{pagetimeQ2Bx7}{(Q2Bx7) If you could exclude companies who are involved in certain activities from your investments on ethical grounds, how important would it be to you to exclude the following activities\link{pipe: hidxPipeQ2Bx5}?}
\item{pagetimeQ2Bx8}{(Q2Bx8) If you could exclude companies who are involved in certain governance activities from your investments, how important would it be to you to exclude the following activities\link{pipe: hidxPipeQ2Bx8}?}
\item{pagetimeQ2Bx9}{(Q2Bx9) If you want to exclude/avoid companies which generate revenues with the activities which are important to you, how strict do you want the exclusion to be?}
\item{pagetimeQ2Bx10}{(Q2Bx10) You want to invest in a way that takes account of social, environmental and governance factors with the objective to improve financial performance. There are different strategies to reduce risks or increase opportunities related to social, environmental or governance factors. Please select the options which you find most promising to achieve this.}
\item{pagetimeQ2Cx1}{(Q2Cx1) You have decided to invest sustainably. If you cannot find a financial product that matches the exact sustainability objectives you have identified, what would best describe your reaction?}
\item{pagetimeQ2Cx2}{(Q2Cx2) Please rate the following statement: I would be willing to accept a financial sacrifice of a certain amount of my annual return for a financial product that closely matched the sustainability objectives I have identified.}
\item{pagetimeQ2Cx3xB}{(Q2Cx3xB) By accepting a financial sacrifice up to a certain amount of your savings, you earn not only a financial return, but also deliver true impact and get evidence that your objectives are achieved in the real economy. There are estimates that this evidence is stronger, the higher the financial sacrifice. In the slider below, you can choose between 0 and 20\% of your final savings. Given your saving behaviour, you would save $ {Pipe_VALUE_X.val} on average after 10 years with 4\% interest. In this case, 1\% sacrifice would mean giving up on $ {Pipe_VALUE_Y.val} and 20\% would mean giving up $ {Pipe_VALUE_Z.val} for achieving impact in the real economy. Please select how much you would be willing to sacrifice at most:}
\item{pagetimeQ2Cx3xA}{(Q2Cx3xA) Generating impact on your sustainability objectives in the real economy is hard to prove. The most evidence for impact might be expected from active management strategies. In the slider below, you can choose between 0 and 20\% of your final savings. Given your saving behaviour, you would save $ {Pipe_VALUE_X.val} on average after 10 years with 4\% interest. In this case, 1\% sacrifice would mean giving up on $ {Pipe_VALUE_Y.val} and 20\% would mean giving up $ {Pipe_VALUE_Z.val} for an active management strategy. Please select how much you would be willing to sacrifice at most:}
\item{pagetimeSection3}{(Section3) Comment element}
\item{pagetimeinfFproduct1}{(infFproduct1) Comment element}
\item{pagetimeQ3x1}{(Q3x1) $ {res.Financialproduct1} Please choose the three words that best describe this marketing claim for you.}
\item{pagetimeQ3x2}{(Q3x2) $ {res.Financialproduct1} Which of the following sentences best describes your understanding of the environmental characteristics associated with this financial product?}
\item{pagetimeQ3x3}{(Q3x3) $ {res.Financialproduct1} Actually, the Equity Fund is invested in companies that pollute less than their competitors. The Equity Fund calculates the difference between the CO2 emissions of these companies and the average of the market every year. There is no evidence that investing more in the Equity Fund will change the operations of these companies though. So it cannot be said that “the more money invested in the fund, the more CO2 reduction activities are developed, the greater your environmental impact.” In the light of this explanation, how would you characterize the marketing claim for this financial product?}
\item{pagetimeinfFproduct2}{(infFproduct2) Comment element}
\item{pagetimeQ3x7}{(Q3x7) $ {res.Financialproduct2} Please choose the three words that best describe this marketing claim for you.}
\item{pagetimeQ3x8}{(Q3x8) $ {res.Financialproduct2} Which of the following sentences best describes your understanding of the environmental characteristics associated with this financial product?}
\item{pagetimeQ3x9}{(Q3x9) $ {res.Financialproduct2} Actually, the Green Equity Fund is invested in the stocks of companies whose production plans are aligned with a 2 °C climate transition pathway. There is no evidence that investing in this fund will contribute to lowering companies’ emissions though. So, it cannot be said that “the more money invested in the fund, the more CO2 emissions are reduced in the world, the lower the increase in global temperatures.” In the light of this explanation, how would you characterize the marketing claim for this financial product?}
\item{pagetimeinfFproduct3}{(infFproduct3) Comment element}
\item{pagetimeQ3x10}{(Q3x10) $ {res.Financialproduct3} Please choose the three words that best describe this marketing claim for you.}
\item{pagetimeQ3x11}{(Q3x11) $ {res.Financialproduct3} Which of the following sentences best describes your understanding of the environmental characteristics associated with this financial product?}
\item{pagetimeQ3x12}{(Q3x12) $ {res.Financialproduct3} Actually, the Green Bond Fund is invested is in the bonds of companies regardless of the nature of their activities (e.g. active in coal-power plants, gas-guzzling cars as well as renewable power, electric car). The Green Bond Fund only invests in the “green” element of these bonds, and the amount available in this green element matches the total amount of money that the companies have invested in environmentally-friendly activities. There is no evidence that your money will increase the number of environmentally-friendly activities though. So it cannot be said that “the more money invested in the fund, the more CO2 reduction activities are developed, the greater your environmental impact.” In the light of this explanation, how would you characterize the marketing claim for this financial product?}
\item{pagetimeinfFproduct4}{(infFproduct4) Comment element}
\item{pagetimeQ3x13}{(Q3x13) $ {res.Financialproduct4} Please choose the three words that best describe this marketing claim for you.}
\item{pagetimeQ3x14}{(Q3x14) $ {res.Financialproduct4} It is not, to date, possible to measure the impact of a financial product on CO2 emissions. All claims containing emission figures are thus misleading. The above claim is an example of a non-misleading claim: Various actions are deployed to maximise the chances that the product triggers CO2 emission reductions, the financial institution communicates on the actions, but no quantified impact is claimed. In the light of this explanation, how would you characterize the marketing claim for this financial product?}
\item{pagetimeSection3EEI}{(Section3EEI) Comment element}
\item{pagetimeQ4x1}{(Q4x1) To diversify investments, most Energy Efficiency Investments EEI and Sustainable Energy Investments SEI funds also invest in other activities than EE and SE. This is common to reduce the risk of the product. Please rank the following statements with regards to this information:}
\item{pagetimeQ4x2}{(Q4x2) Comment element}
\item{pagetimeINVEST1}{(INVEST1) Comment element}
\item{pagetimeINVEST2}{(INVEST2) Comment element}
\item{pagetimeINVEST3}{(INVEST3) Comment element}
\item{pagetimeINVEST4}{(INVEST4) Comment element}
\item{pagetimeINVEST5}{(INVEST5) Comment element}
\item{pagetimeQ4x3}{(Q4x3) Please grade, for each investment (from investment 1 to 5), the following statement: I would be interested to undertake investments in the described investment options:}
\item{qtime}{ Total Interview Time}
}
}
\usage{
ESG_france_2021
}
\description{
The survey aimed to retail investor sustainability preferences and their understanding of marketing claims. The survey is structured in 9 parts : A Sociodemographic information, B Financial information, C Sustainability profile, D Impact goals, E Evidence preferences, F Exclusion topics and scope, G Willigness for tradeoffs, H Misleading claims, I Believes about EEI/SEI products.
}
\keyword{datasets}
