% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mystery_EEI_2023.R
\docType{data}
\name{Mystery_EEI_2023}
\alias{Mystery_EEI_2023}
\title{Mystery Shopping campaign 2023}
\format{
A tibble with 98 rows and 42 variables:

\describe{

\item{ID}{Evaluation_ID}
\item{Client_Survey_Name}{Client_Survey_Name}
\item{Evaluation_Date}{Evaluation_Date}
\item{Location_ID}{Location_ID}
\item{Location_Name}{Location_Name}
\item{Country}{Country}
\item{Evaluation_Status}{Evaluation_Status}
\item{q1}{1. What is your Mystery Shopper Profile?}
\item{q2}{2. Date of appointment}
\item{q3}{3. What is the name of the bank where you had an appointment? Please select.}
\item{q4}{4. What is the title of the person you had an appointment with?}
\item{q5}{5. What was the gender of the advisor?}
\item{q6}{6. How long did the appointment last?}
\item{q7}{7. Did the advisor bring up the subject of sustainability preferences without any prompting?}
\item{q8}{8. Did the advisor focus first on assessing your financial/investment objectives, before starting to talk about sustainability preferences?}
\item{q9}{9. Did the advisor give you an explanation of sustainability preferences?}
\item{q10}{10. What was included in the advisor's explanation of sustainability preferences? (Pls select all that apply)}
\item{q11}{11. In the explanation of sustainability preferences, did the advisor refer to different sustainability goals/objectives you might have for your investments?     (Pls select all that apply)}
\item{q12}{12. Was the explanation linked to the financial products which the advisor could recommend or to all financial products on the market?}
\item{q13}{13. Was the explanation of sustainability preferences provided in a neutral and unbiased manner?}
\item{q14}{14. When you asked whether investing in green products was risky, how did the advisor respond?}
\item{q15}{15. Did you fully understand the explanation of sustainability preferences?}
\item{q16}{16. Did the advisor check that you had understood the explanation of sustainability preferences?}
\item{q17}{17. Did the advisor appear well trained and competent in sustainable finance generally?}
\item{q18}{18. If you noticed a problematic lack of knowledge/expertise at any moment of the interview, please indicate those moments:}
\item{q19}{19. Did the advisor ask what your own sustainability preferences are?}
\item{q20}{20. Did the advisor ask you about your own sustainability preferences before presenting any financial products to you?}
\item{q21}{21. Did the advisor ask you about the minimum proportion you want to invest in line with your sustainability preferences?}
\item{q22}{22. How was the minimum proportion you want to invest in line with your sustainability preferences assessed?}
\item{q23}{23. When you asked whether it can be guaranteed that human rights are respected by companies which are invested in, how did the advisor respond?}
\item{qA1}{A1. When you expressed a desire to have a minimum proportion of 100\% invested in environmentally sustainable investment with a focus on climate mitigation and climate adaptation, how did the advisor respond?}
\item{qB1}{B1. When you expressed a desire to invest in a way which generates positive impact in the real world and asked what could prove/demonstrate the impact of any investment, how did the advisor respond?}
\item{q24}{24. What tools did the advisor use to record information about your sustainability preferences?}
\item{q25}{25. After asking all these questions, did the advisor summarize your profile in relation to sustainability preferences?}
\item{q26}{26. Did this summary correspond with your expectations/needs?}
\item{q27}{27. Did the advisor present a financial product to you without asking whether you want to adapt your sustainability preferences?}
\item{q30}{30. If the advisor was not able to present a financial product which matches your sustainability preferences, what were the advisor's next steps?}
\item{q31}{31. If the advisor asked you whether you wanted to adapt your sustainability preferences because they do not have a suitable financial product, was the advisor transparent that other financial products which do match your sustainability preferences might be available from other advisors?}
\item{q32}{32. If the advisor asked you whether you wanted to adapt your sustainability preferences, did you feel you were influenced to adapt your sustainability preferences to a financial product which the advisor could recommend? Did the advisor put you under pressure to choose a product?}
\item{q33}{33. What tools did the advisor use to record information about whether you changed your sustainability preferences?}
\item{q34}{34. At the end of the appointment, did the advisor give you a summary?}
\item{q35}{35. In case the advisor did not give you a summary at the end of the appointment, after you asked for it, did you receive one?}
}
}
\usage{
Mystery_EEI_2023
}
\description{
{Description}{The focus of the this mystery shopping campaign was to assess the compliance with the new MiFID II regulation and broader sustainability motivations. We allocated 2 hypothetical investor profiles to the mystery shoppers including. We assessed the following topics (please note that for duty of confidentiality we removed personal information of mystery shoppers and the bank name and recommended products): - Explanation of sustainability preferences - Advisor knowledge and expertise - Identification of sustainability preferences - Financial product presentation/recommendation - Final conclusion of the meeting.}
}
\details{
{Usage}{Mystery_EEI_2023}
}
\keyword{datasets}
